last refactor I beg
